Moved Membership from Lis\v2 to Membership
